Changed floats to doubles, fixed warnings about int to size_t conversion

+++ source/FluidSim/FluidSim/App/Fluid3D.cpp
@@ -6,7 +6,7 @@
 
 Fluid3D::Fluid3D(const ApplicationData& inData)
 {
-	float start = glfwGetTime();
+	double start = glfwGetTime();
 	std::cout << "Initializing particles: ";
 	// Initialize particles
 	mParticles.reserve(inData.GetNumParticles());
